+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## To be released
+### Features
+* Add option to `widgets::List` to repeat the hightlight symbol for each line of multi-line items.
 
 ## v0.16.0 - 2021-08-01